[mdf][all] Мир стал целочисленным
Дробных величин в состоянии мира больше нет ни одной. Причина не во вкусе
к целым числам: дробь в состоянии всегда тянула за собой второе поле —
накопитель, в котором дробь дожидалась, пока накопится до целого. Так в
мире и завелись mineralPending, proteinPending, dungPending, humus.pending
— поля, не означавшие ничего, кроме "0.004 за тик не помещается в штуку".
Целая шкала убирает и дробь, и накопитель разом.

Шкала одна на весь проект (новый core/Scale.hpp): доли — в тысячных
(kFull = 1000), высоты и глубины — в тысячных прежней единицы, счётчики
тела — там же. Времена, расстояния и крупицы целыми были и раньше.

Скорости, которые целыми не получаются, выражены двумя способами, и оба
обходятся без накопителя:
  - СРОК вместо скорости: испарение воды, разложение перегноя, навоз.
    "Одна тысячная раз в пять тиков" — то же самое, что "две
    десятитысячных за тик", только целое.
  - РАЗНОСТЬ ДВУХ ДЕЛЕНИЙ: взросление животного выводится из возраста,
    который тело и так хранит.
Накопитель остался ровно там, где скорость модулируется переменным и из
возраста не выводится: рост растения (скорость * влага) и шаг животного.
Оба целые остатки, а не дроби.

Течение воды: разлив по перекрывающимся девяткам клеток делит правку на
девять, и в целых числах это теряло бы до восьми тысячных глубины на
клетке каждый тик — вчетверо больше, чем испаряется, то есть карта
высыхала бы сама собой без причины. Неделящийся остаток теперь
переносится на следующий тик (WaterComponent::flowRemainder), и объём
воды сохраняется точно. Заодно исчез порог kWaterMinDepth: на целой шкале
меньше единицы глубины не бывает, и ноль сам означает "воды нет".

Гены тоже целые: Trait ссылается на int-поле, а расклад бюджета
преимуществ внутри Genetics.hpp остаётся дробным — он не состояние мира,
а способ вывести гены один раз при рождении. Дробное становится целым на
границе, в genomeOf, с округлением к ближайшему.

Протокол: слои уходят ровно теми числами, что лежат в компонентах. Поля
"scale" и перевода в тысячные (encodeMilli/decodeScaled с переменным
делителем) больше нет — делить не на что.

Панель генерации: floatRow удалена, дробных ползунков не осталось ни
одного. noise_frequency (0.02) стала feature_size (50 тайлов) — обратная
величина, зато читаемая глазом.

Численность поголовья заметно выросла против прежней: срез платных черт
освободил бюджет, а трава без функции пригодности растёт быстрее. Мир от
этого не сломался — на трёх прогонах по 12000 тиков и трава, и стадо, и
хищники живут, качаясь циклами; хищники теперь доживают до конца прогона,
а не вымирают к третьей тысяче, как раньше.

+++ core/include/core/components/DesireComponent.hpp
@@ -79,7 +79,8 @@ inline Desire desireFromName(const std::string& name) {
 // Остаётся то, что хранить обязано: mating (в теле его не прочитать) и
 // current (у желания есть инерция).
 struct DesireComponent {
-    float mating = 0.0f;
+    // Желание пары, 0..kFull (core/Scale.hpp).
+    int mating = 0;
     Desire current = Desire::Idle;
 };
